requirements.txt
requests>=2.25.0
Pillow>=9.0.0
pillow-heif>=0.10.0